+++ src/main/kotlin/com/nylas/models/CreateDraftRequest.kt
@@ -72,6 +72,12 @@ data class CreateDraftRequest(
    */
   @Json(name = "custom_headers")
   val customHeaders: List<CustomHeader>? = null,
+  /**
+   * When true, the message body is sent as plain text and the MIME data doesn't include the HTML version of the message.
+   * When false, the message body is sent as HTML.
+   */
+  @Json(name = "is_plaintext")
+  val isPlaintext: Boolean? = null,
 ) : IMessageAttachmentRequest {
   /**
    * Builder for [CreateDraftRequest].
@@ -90,6 +96,7 @@ data class CreateDraftRequest(
     private var replyToMessageId: String? = null
     private var trackingOptions: TrackingOptions? = null
     private var customHeaders: List<CustomHeader>? = null
+    private var isPlaintext: Boolean? = null
 
     /**
      * Sets the from address.
@@ -184,8 +191,17 @@ data class CreateDraftRequest(
     fun customHeaders(customHeaders: List<CustomHeader>?) = apply { this.customHeaders = customHeaders }
 
     /**
-     * Builds a [SendMessageRequest] instance.
-     * @return The [SendMessageRequest] instance.
+     * Sets whether the message body is sent as plain text.
+     * When true, the message body is sent as plain text and the MIME data doesn't include the HTML version of the message.
+     * When false, the message body is sent as HTML.
+     * @param isPlaintext Whether the message body is sent as plain text.
+     * @return The builder.
+     */
+    fun isPlaintext(isPlaintext: Boolean?) = apply { this.isPlaintext = isPlaintext }
+
+    /**
+     * Builds a [CreateDraftRequest] instance.
+     * @return The [CreateDraftRequest] instance.
      */
     fun build() =
       CreateDraftRequest(
@@ -202,6 +218,7 @@ data class CreateDraftRequest(
         replyToMessageId,
         trackingOptions,
         customHeaders,
+        isPlaintext,
       )
   }
 }